Minor Version: 0.2.0
In this minor release Breakfast gets the status bar. In either the top
or bottom left part of the screen will be a small bar which outputs the
latest build status.

In a later release planning to make it more configurable and full
featured (toggle abilities for reload strategies, custom CSS, more
locations). But for now it will suffice.

Bug Fixes:
  - Setup bug if user does not have Brunch installed globally
    Brunch spawner was using the global version of brunch instead of
    using the binary in the node_modules/ directory.