bug: Fix potential Heroku Regex Bug
Slight error in the regular expression that determines whether or not an
asset is already fingerprinted or not. The expected behavior is that any
asset matching FINGERRINT_REGEX, that is a filename, followed by a dash
and then 32 hex characters, is not digested again.

Example filename: filename-869269cdf1773ff0dec91bafb37310ea.js

It appears that on some deployments Heroku will generate folder names
which happen to match the FINGERPRINT_REGEX.

The Regex was meant to look for a dot after the hex string. However, the
dot is unescaped - meaning that any character after the 32 hex string
would get matched. Including folder slashes. So paths like:

folder-869269cdf1773ff0dec91bafb37310ea/app.js

Would be skipped over and not digested by Breakfast. Then the clean task
seeing nothing in the Breakfast Manifest would remove every file in the
output directory. So the app would have no styles.